requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
streamlit